+++ src/ui.ts
@@ -151,11 +151,14 @@ const createMap = (container: string | HTMLElement, options?: any): Promise<any>
 
     onDirectionQueryWillBeSent: (query: any): any => query,
     onDirectionWillBeDisplayed: (direction: any, directionOptions: any): any => ({ direction, options: directionOptions }),
+
     onElementWillBeSelected: (element: any, opts: any): any => opts,
+
     onFollowButtonClickWithoutLocation: (): void => null,
     onInformationButtonClick: (): void => null,
-    onSearchQueryWillBeSent: (searchString: string, searchOptions: any): any => ({ searchString, searchOptions }),
-    onSearchResultWillBeDisplayed: (results: any): any => results,
+    onReceiveSearchResults: (results: any): any => results,
+    onSearchQueryWillBeSent: (searchString: string, searchOptions: any, focusedField: string): any => ({ searchString, searchOptions }),
+    onSearchResultWillBeDisplayed: (template: string, templateOptions: any): any => ({ template, options: templateOptions }),
     onSelectedChange: (): void => null,
     shouldShowInformationButtonFor: (element: any): boolean => false,